fix(ui): split renderer fixtures
