+++ npm/ui/core/scripts/check-renderers.ts
@@ -4,6 +4,8 @@ import path from "node:path";
 
 import { compileSfc, type SfcCompileOptionsNapi } from "@vizejs/native";
 
+import { rendererFixtures } from "./renderer-fixtures.ts";
+
 interface RendererLane {
   /** Stable lane name emitted in CI diagnostics. */
   readonly name: "dom" | "ssr" | "vapor";
@@ -18,45 +20,7 @@ const rendererLanes: readonly RendererLane[] = [
 ];
 
 const inlineFixtures = [
-  {
-    filename: "InertOutsideConsumer.vue",
-    source: String.raw`<script setup lang="ts">
-import { ref } from "vue";
-import { useInertOutside } from "./inert-outside.ts";
-
-const root = ref<HTMLElement | null>(null);
-const isolation = useInertOutside({ root, mode: "both" });
-</script>
-
-<template>
-  <div ref="root" :data-active="isolation.isActive.value || undefined">
-    Modal content
-  </div>
-</template>
-`,
-  },
-  {
-    filename: "FocusScopeConsumer.vue",
-    source: String.raw`<script setup lang="ts">
-import { ref } from "vue";
-import { useFocusScope } from "./focus-scope.ts";
-
-const root = ref<HTMLElement | null>(null);
-const scope = useFocusScope({
-  root,
-  contain: true,
-  autoFocus: true,
-  restoreFocus: true,
-});
-</script>
-
-<template>
-  <div ref="root" :data-active="scope.isActive.value || undefined">
-    <button type="button">Inside</button>
-  </div>
-</template>
-`,
-  },
+  ...rendererFixtures,
   {
     filename: "SpatialNavigationConsumer.vue",
     source: String.raw`<script setup lang="ts">
